rename tictactoe strategy file
